Changed spec functions to take a pow as a parameter instead of just grabbing from the constant buffer.  This allows me to fetch a pow from vertex or look up table or whateva.
And for that, I added a power 4x4.  I wanted a float array but float
arrays get very odd in constant buffers.  They pad out to float4, so I
just use array notation with a matrix.  It goofs up the define, but I'll
fix it later.  Somehow.

Might add an emissive table for glowing eyes and such later.

Had a bit of trouble with mixing textures and vert colors.  I made a
texture just to test the shader, and it is mostly empty, just had rosy
skin for the knees and cheeks.

When using cel shading they sometimes glow a bit in super dark areas, or
white out in bright.  I'm not sure how to deal with that or what the
proper way to combine is.  Add makes them sort of glowy, but multiply
looks very off.

c * alpha + other * (1 - alpha) looked strange too.  I'll fix it later
when I have more stuff in the texture.  Not even sure I really need a
texture.  This stuff is going to be pixelated and viewed top down most
of the time.